[test] Split out migration pattern tests
